+++ HISTORY.md
@@ -1,5 +1,9 @@
 # history
 
+## 0.3.12 / 2019-05-20
+* `FIXED` when required is set, jsxrules will auto add not empty rule
+* `NEW` add new prop requiredErrMsg for required prop
+
 ## 0.3.11 / 2019-04-12
 
 * `CHANGED` remove errMsg type check